+++ src/UuidFixtures.php
@@ -1,10 +1,12 @@
 <?php
 
-/**
- * xAPI statement activity fixtures.
+/*
+ * This file is part of the xAPI package.
  *
- * These fixtures are borrowed from the
- * {@link https://github.com/adlnet/xAPI_LRS_Test Experience API Learning Record Store Conformance Test} package.
+ * (c) Christian Flothmann <christian.flothmann@xabbuh.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
  */
 
 namespace Xabbuh\XApi\DataFixtures;